Rename control map file to reflect actual behavior
